requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
async-timeout==4.0.3
attrs==24.2.0
Automat==24.8.1
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
constantly==23.10.4
cryptography==43.0.1
cssselect==1.2.0
dataclasses-json==0.6.7
datasets==3.1.0
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
emoji==2.14.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
filelock==3.16.1
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.9.0
gotrue==2.9.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.2
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jq==1.8.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-ollama==0.2.0
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.147
lxml==5.3.0
Markdown==3.7
markdownify==0.13.1
marshmallow==3.23.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
ollama==0.3.3
openai==1.52.0
orjson==3.10.9
packaging==24.1
pandas==2.2.3
parsel==1.9.1
postgrest==0.17.1
propcache==0.2.0
Protego==0.3.1
psutil==6.1.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyDispatcher==2.0.7
PyMuPDF==1.24.11
pymupdf4llm==0.0.17
PyMuPDFPro==1.24.11
pyOpenSSL==24.2.1
pypdf==5.0.1
pysbd==0.3.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
queuelib==1.7.0
ragas==0.2.6
RapidFuzz==3.10.0
realtime==2.0.5
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rouge_score==0.1.2
Scrapy==2.11.2
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
storage3==0.8.1
StrEnum==0.4.15
supabase==2.9.0
supafunc==0.6.1
tenacity==9.0.0
tiktoken==0.8.0
tldextract==5.1.2
tomli==2.0.2
tqdm==4.66.5
Twisted==24.7.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.1
unstructured-client==0.26.1
urllib3==2.2.3
w3lib==2.2.1
webencodings==0.5.1
websockets==13.1
wrapt==1.16.0
xxhash==3.5.0
yarl==1.15.3
zope.interface==7.1.0
